TODO消化

+++ src/p0009.rs
@@ -22,7 +22,7 @@ fn solve() -> i64 {
         .flat_map(|m| {
             let start = if m % 2 == 0 { 1 } else { 2 };
             (start..m).step_by(2).filter_map(move |n| {
-                if (m - n) % 2 == 1 && gcd(m as u64, n as u64) == 1 {
+                if (m - n) % 2 == 1 && gcd(m, n) == 1 {
                     Some((m * m - n * n, 2 * m * n, m * m + n * n))
                 } else {
                     None
